loader: implement new semantics for entity overriding
This implements the new scheme for allowing entities to
be overridden. If an XML file relative path matches one
in an earlier datbase directory, the original file is
not loaded at all. Any files in a ENTITY-NAME.d will
now augment earlier data rather than replace it.

eg consider these files

1. /usr/share/libosinfo/db/os/fedoraproject.org/fedora-22.xml
2. /etc/libosinfo/db/os/fedoraproject.org/fedora-22.xml
3. $HOME/.config/libosinfo/db/os/fedoraproject.org/fedora-22.d/extra.xml

File #1 will not get loaded, because file #2 is detected
as a full override. File #2 will get loaded to define the
main Fedora 22 OS, and then file #3 will get loaded to
augment the main data.

Most of the PCI and USB device info in the local DB is
just augmenting the pci.ids and usb.ids master files,
so these need to move to .d directories now

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>