Fix build system after creating initial repository
The initial repository state was created from libosinfo
GIT at revision:

  commit 501541a336cd29d3da48584f022ce329fef418dd
  Author: Fabiano Fidêncio <fidencio@redhat.com>
  Date:   Thu May 12 18:54:58 2016 +0200

    win,installer: Fix "Windwos" typo

It was initialized using the following commands:

 $ git clone libosinfo osinfo-db-tools
 $ cd osinfo-db-tools
 $ git filter-branch --index-filter 'git rm -r --cached \
      --ignore-unmatch src examples data db test \
      docs/reference tools/osinfo-detect.c \
      tools/osinfo-install-script.c \
      tools/osinfo-query.c po osinfo' HEAD
 $ git filter-branch -f --prune-empty HEAD
 $ git for-each-ref --format="%(refname)" refs/original/ | xargs -n 1 git update-ref -d
 $ git reflog expire --expire=now --all
 $ git gc --prune=now

This patch now edits the various build system files to
remove references to bits of the tree that were deleted,
and rename from libosinfo to osinfo-db-tools.

Currently this module contains a single command

  osinfo-db-validate

It will later be augmented with tools for importing
and exporting database archives.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>