Add osinfo-db-export tool for packing an osinfo database archive
Add a new osinfo-db-export tool which creates an archive in the
tar (pax) format, with XZ compression and can filling it with the
database files from one of the standard libosinfo database locations,
or a custom directory.

This makes it easy for users to create their own local custom database
content and then distribute and import it on many machines. It will
also be used for creating official release archives of the libosinfo
database from upstream.

 $ osinfo-db-export --verbose  osinfo-db-2016-06-15.tar.xz
 osinfo-db-export: d osinfo-db-2016-06-15/
 osinfo-db-export: d osinfo-db-2016-06-15/datamaps
 osinfo-db-export: r osinfo-db-2016-06-15/datamaps/windows-lang.xml
 osinfo-db-export: r osinfo-db-2016-06-15/datamaps/x11-keyboard.xml
 osinfo-db-export: d osinfo-db-2016-06-15/devices
 osinfo-db-export: r osinfo-db-2016-06-15/devices/qemu-pci.xml
 osinfo-db-export: r osinfo-db-2016-06-15/devices/qemu-ps2.xml
 osinfo-db-export: r osinfo-db-2016-06-15/devices/qemu-usb.xml
 osinfo-db-export: r osinfo-db-2016-06-15/devices/virtio-pci.xml
 osinfo-db-export: r osinfo-db-2016-06-15/devices/xen.xml
 osinfo-db-export: d osinfo-db-2016-06-15/hypervisors
 osinfo-db-export: r osinfo-db-2016-06-15/hypervisors/kvm.xml
 ...snip...

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>