Fix build with clang
The osinfo_db_import_create_reg() function implements error handling
via jumping to cleanup and returning, however, this way the return
value is not initialised. This breaks build with clang and probably
could lead to an undefined behavior for users of this function.

Fix by initialising ret with -1 by default.