Don't expand entities when parsing XML
The XML_PARSE_NOENT flag to libxml will cause it to expand all entities in the
input XML document when parsing. Doing this is bad practice if the XML input
file comes from an untrusted source, because it can cause the XML parser to load
arbitrary files that are readable by the user running XML parsing. This is does
not have an security consequences given the scenario in which osinfo-db-validate
is run since the intended usage is to validate files that are written by the
local user, or by the upstream libosinfo maintainers.

In the future though, libosinfo might be able to dynamically download data from
the website to refresh its local database, so it is wise to avoid entity
expansion as a hardening step.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>