osinfo-db-import: fix loading of local files
When the argument can be either a URI or a local path, it the
g_file_new_for_commandline_arg constructor must be used. In
addition, when the GFile is a native path, we must not use the original
path, but instead call g_file_get_path(), so that file:/// URIs are
correctly handled by stripping the protocol.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>