tests: Add osinfo-db-{export,import} tests
Add a set of functional tests to ensure osinfo-db-export and
osinfo-db-import work as expected. The covered scenearios are:
- osinfo-db-export DIR
  - It's a negative test
- osinfo-db-export --system and osinfo-db-import --system
- osinfo-db-export --local FILENAME
- osinfo-db-import --local FILENAME
- cat FILENAME | osinfo-db-import -
- cat FILENAME | osinfo-db-import
- osinfo-db-export --user --license LICENSE --version VERSION
  VERSIONED_FILENAME
- osinfo-db-import --root / --user VERSIONED_FILENAME
- osinfo-db-import --user URL
- osinfo-db-import --latest

The last two tests are only enabled when OSINFO_DB_TOOLS_NETWORK_TESTS
is set.

Last but not least, differently than the tests for osinfo-db-path and
osinfo-db-validate, those tests pretty much only make sense when taken
as pairs. So, we export a db, import it back and ensure the content is
the same.

https://gitlab.com/libosinfo/osinfo-db-tools/issues/2

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>