spec: Add %check target (and the needed deps)
Now that we have tests as part of osinfo-db-tools, let's enable them
when building an RPM.

A new set of BuildRequires dependencies have been added as our tests
depend on python3, python3-pytest, and python3-requests.

The %check target has been added unconditionally the tests are only
executed in case the deps are found in the system.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>