Drop autotools support
Let's try to drop autotools support as meson is in place. There are a
few files that were not dropped though, as:
- cfg.mk
- maint.mk
- GNUmakefile
- build-aux/bracket-spacing.pl
- build-aux/gitlog-to-changelog
- build-aux/useless-if-before-free
- build-aux/vc-list-files

From the list above, build-aux/gitlog-to-changelog is currently used by
meson as part of the `dist`.

All the other files, with a few changes, will be useful for meson as,
with a few changes, we will be able have `ninja syntax-check` working.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>