import: Use libsoup for http:// & https:// requests
As osinfo-db-import handles non-local requests, it'd end up relying on
GVFS under the hood, which would cause the app to not work when running
as root.

In order to avoid this situation, let's rely on libsoup for this case.

It's important to mention that glib minimum version has been bumped to
2.44 as that's the version where g_input_stream_read_all() has been
introduced.

https://gitlab.com/libosinfo/libosinfo/issues/30

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>