Remove AUTHORS, MAINTAINERS, and ChangeLog* files
There's no reason for keeping generating and distributing AUTHORS,
MAINTAINERS, and Changelog files.

AUTHORS is explicitly required by autoconf, which is not used anymore.

MAINTAINERS is something that does not make much sense to have, as the
preferred way for people to report issues should be the one described in
libosinfo.org and never ever contacting a maintainer directly.

ChangeLog is also something that does not make much sense to have, as
`git log` can provide all the info needed, in case someone wants to dig
into the project's history.

*: It's important to mention that we never had any "ChangeLog" file, the
file was generated by gitlog-to-changelog script, which was also removed
and then put into the distributed tarball.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>