gitlab: Enable ccache usage
The $PATH definition has always included a reference to the symlink
farm that is used to enable transparent ccache usage, but since the
directory itself does not exist in the container image it never did
anything.

Since libvirt-jenkins-ci commit 2563aebb6c5c, container images do
actually contain the symlink directory, so it's finally possible to
enable ccache usage. This requires including the correct directory
in $PATH, as well as telling 1) ccache where to store its data and
2) GitLab that this data should be preserved across jobs.

The exact cache size (50 MB) is sort of arbitrary, but it should
be enough to store build artifacts for both osinfo-db-tools and
libosinfo for the foreseeable future.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>