test_osinfo_db_export_import_system: fix on day change
In case the current day changes while osinfo-db-export is run, then
the test does not find the generated archive anymore: this is because
the filename of the generated archive has the timestamp of the date.

As workaround, build the filename for today and tomorrow, checking that
one of them must exist.

Fixes: #5

Signed-off-by: Pino Toscano <ptoscano@redhat.com>