osinfo-db-import: Make osinfo_db_get_latest_info more generic
Future patches will add support for importing unreleased nightly builds
of osinfo-db which would use much of the same code path as '--latest'
already does with regards to fetching information about the resulting
archive URL to download.
This patch makes the existing function more generic by including the
URL to the JSON to be parsed and renaming it to osinfo_db_get_info.

Signed-off-by: Erik Skultety <eskultet@redhat.com>