osinfo-db-import: Only poll the 'version' attribute from latest.json
Previous introduced a more generic osinfo_db_get_info version that
should suit both the '--latest' as well as '--nightly' scenarios.
But before a caller function for the latter is introduced, make sure
that the 'version' metadata JSON attribute is queried only from
latest.json and not nightly.json, because the latter doesn't, quite
understandably, expose it.

Signed-off-by: Erik Skultety <eskultet@redhat.com>