osinfo-db-import: Rename latest_url to archive_url
Future patch will introduce support for importing nightly build
archives which will obviously have a different URL. The way this is
currently reflected in the code is by overriding the 'archive' variable
with the proper URL which however, is named 'latest_url'. Rename it to
a more generic name that can be re-used for the other scenario as well.

Signed-off-by: Erik Skultety <eskultet@redhat.com>